+++ NEWS.md
@@ -1,9 +1,20 @@
-# omopgenerics 1.1.2
+# omopgenerics 1.3.0
 
-* remove NA estimates in transformToSummarisedResult by @catalamarti in #702
+* write method fro summary.cdm_source by @catalamarti in #719 #720
+* Add query id in logging files by @catalamarti in #716
+* Expanding omopgenerics vignette by @catalamarti in #721
+* Indexes experimental functions by @catalamarti in #722 #723 #724
+
+# omopgenerics 1.2.0
+
+* Remove NA in estimates in transformToSummarisedResult by @catalamarti in #702
 * Create logging functions by @catalamarti in #700
 * Allow strata to be a character by @catalamarti in #703
 * Remove settings that are NA after filterSettings by @catalamarti in #704
+* `validateWindowArgument` force snake_case names by @catalamarti in #711
+* Keep cohort_table class after collect by @catalamarti in #710
+* `dplyr::as_tibble` for codelist by @catalamarti in #712
+* `type` -> `codelist_type` by @catalamarti in #709
 
 # omopgenerics 1.1.1